# doc/EZshopBackEndApplication_swagger.yaml
openapi: 3.0.3
info:
  title: EZShop Back End Application API
  version: 1.0.0
  description: 'OpenAPI definition for the EZShop back end application.


    Assumptions:

    - Backend implemented in Python (e.g., FastAPI/Flask), stateless JWT auth.

    - Entities inferred from EZShopInterface.java: User, ProductType, Customer, Order,
    SaleTransaction, ReturnTransaction, BalanceOperation, TicketEntry, LoyaltyCard.

    - Some fields are inferred; adjust to match your domain models.

    '
servers:
- url: http://127.0.0.1:8000/api/v1
  description: Local dev
- url: https://api.ezshop.example.com
  description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserType:
      type: string
      enum:
        - Administrator
        - Cashier
        - ShopManager
      description: User type in the system
      example: "Administrator"
...
    UserCreateRequest:
      type: object
      required:
        - username
        - password
        - type
      properties:
        username:
          type: string
          minLength: 5
          description: Unique username for the user
          example: "admin_user"
        password:
          type: string
          minLength: 5
          format: password
          description: User password (mandatory field)
          example: "SecurePass123!"
        type:
          $ref: '#/components/schemas/UserType'
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 5
          description: New username for the user
          example: "updated_username"
        password:
          type: string
          minLength: 5
          format: password
          description: New password for the user
          example: "NewSecurePass789!"
        type:
          $ref: '#/components/schemas/UserType'
    UserResponse:
      type: object
      required:
        - id
        - username
        - type
      properties:
        id:
          type: integer
          description: Unique user ID
          example: 1
        username:
          type: string
          description: Username
          example: "admin_user"
        type:
          $ref: '#/components/schemas/UserType'
      description: Complete user representation (without password)
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - name
      properties:
        code: 
          type: integer
          description: HTTP status code of the error
          example : 400
        message:
          type: string
          description: message of the error
          example: Conflict
        name:
          type: string
          description: Descriptive error message
          example: "User not found"
      description: Standard structure for error responses
...
paths:
...
  /users:
    post:
      tags:
        - users
      summary: Create a new user
      description: Creates a new user in the system. Requires authentication as **Administrator**. 
        The success response is the user with its assigned ID.
      operationId: createUser
      security:
        - bearerAuth: [Administrator]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
            examples:
              administrator:
                summary: Create administrator
                value:
                  username: "admin_user"
                  password: "SecurePass123!"
                  type: "Administrator"
              standard_user:
                summary: Create standard user
                value:
                  username: "standard_user"
                  password: "UserPass456!"
                  type: "Standard"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 1
                username: "admin_user"
                type: "Administrator"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_password:
                  summary: Missing password
                  value:
                    detail: "Password is a mandatory field"
                missing_type:
                  summary: Missing user type
                  value:
                    detail: "User type is a mandatory field"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "User with username 'admin_user' already exists"
    get:
      tags:
        - users
      summary: List all users
      description: Retrieves the list of all registered users. Requires authentication as **Administrator**.
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
              example:
                - id: 1
                  username: "admin_user"
                  type: "Administrator"
                - id: 2
                  username: "standard_user"
                  type: "Standard"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{user_id}:
    get:
      tags:
        - users
      summary: Get a specific user
      description: Retrieves the details of a user by ID. Requires authentication as **Administrator**.
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user to retrieve
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 1
                username: "admin_user"
                type: "Administrator"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "User not found"
                name: "NotFoundError"
    put:
      tags:
        - users
      summary: Update a user
      description: Updates the information of an existing user. Requires authentication as **Administrator**. 
        The new username must be unique. This operation can update the UserType and password as well.
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user to update
          schema:
            type: integer
            minimum: 1
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
            example:
              username: "updated_username"
              password: "NewSecurePass789!"
              type: "Administrator"
      responses:
        '201':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 1
                username: "updated_username"
                type: "Administrator"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Bad request"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "User not found"
                name: "NotFoundError"
        '409':
          description: Conflict - Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 409
                message: "User with username 'updated_username' already exists"
                name: "ConflictError"
    delete:
      tags:
        - users
      summary: Delete a user
      description: Deletes a user from the system. Requires authentication as **Administrator**.
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user to delete
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '204':
          description: User deleted successfully (no content)
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "User not found"
...